// src/app/model/item.ts
export class Item {

    id: number;
    fsaCppPurchaseOrderId: number;
    bidNumber: string;
    bidItemCodeId: number;
    itemNumber: number;
    itemMake: string;
    itemType: string;
    itemModelNumber: string;
    itemDescription: string;
    itemAmount: number;
    qty: number;
    adminFeeDue: number;
    fsaFee: number;
    facFee: number;
    ffcaFee: number;
    poIssueDate: Date;
    estimatedDeliveryDate: Date;
    updatedBy: string;
    createdBy: string;

}